refactor(desktop): promote AgentCommentComposer to shared pane components
The browser pane's design mode is its second consumer, so it moves from
DiffPane/components to the panes' shared components directory per the
co-location rules. The line-range header generalizes to a contextLabel
prop (DiffPane passes its line label), the footer row wraps in narrow
panes, and AgentPickerSelect + useDiffCommentTarget become part of the
barrel's public API for composers that build their own chrome.
