+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/AgentCommentComposer/index.ts
@@ -3,3 +3,5 @@ export type {
 	AgentTarget,
 } from "./AgentCommentComposer";
 export { AgentCommentComposer } from "./AgentCommentComposer";
+export { AgentPickerSelect } from "./components/AgentPickerSelect";
+export { useDiffCommentTarget } from "./hooks/useDiffCommentTarget";
